Make Windows compilation happy.

+++ arangod/Cluster/ClusterMethods.cpp
@@ -3314,7 +3314,7 @@ arangodb::Result hotBackupList(std::vector<ServerID> const& dbServers, VPackSlic
       front._nrFiles = totalFiles;
       front._serverId = "";  // makes no sense for whole cluster
       front._isAvailable = i.second.size() == dbServers.size();
-      front._nrPiecesPresent = i.second.size();
+      front._nrPiecesPresent = static_cast<unsigned int>(i.second.size());
       hotBackups.insert(std::make_pair(front._id, front));
     }
   }
